# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Java Backend Challenge API",
    "version": "0.1.0",
    "description": "Sports training scheduling and registration API. All eight operations are implemented. Coach names retain the public name field and normalize ASCII whitespace. Timestamps require an explicit offset, are truncated to microseconds and return UTC. Errors use application/json with stable codes and optional fieldErrors. See README.md for setup and a complete HTTP walkthrough, and documentation/ADR-001-coach-name-expansion.md for database compatibility and rollback decisions."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local service; runtime Swagger documents implemented endpoints."
    }
  ],
...
  "paths": {
    "/api/coaches": {
      "post": {
        "operationId": "createCoach",
        "summary": "Create a coach",
        "tags": [
          "Coaches"
        ],
        "description": "Name and email must be nonblank; email must be valid. Emails are not unique identities. Normalize surrounding and repeated ASCII whitespace (space, tab, line breaks, vertical tab and form feed) in coach names. Internally store the first word as first_name and the remainder as last_name; a single word has an empty last_name. Return the same id/name/email DTO with the normalized public name. Participant names remain unchanged.",
        "responses": {
          "201": {
            "description": "Resource created; body includes its ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coach"
                },
                "example": {
                  "id": "11111111-1111-4111-8111-111111111111",
                  "name": "Alex Rivera",
                  "email": "alex@example.com"
                }
              }
            }
          },
          "400": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "invalidFields": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request contains invalid fields.",
                      "instance": "/api/coaches",
                      "code": "INVALID_REQUEST",
                      "fieldErrors": [
                        {
                          "field": "email",
                          "message": "must be a valid email"
                        }
                      ]
                    }
                  },
                  "malformedJson": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request body must contain valid JSON.",
                      "instance": "/api/coaches",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/coaches",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCoach"
              },
              "example": {
                "name": "  Alex   Rivera  ",
                "email": "alex@example.com"
              }
            }
          }
        }
      }
    },
    "/api/participants": {
      "post": {
        "operationId": "createParticipant",
        "summary": "Create a participant",
        "tags": [
          "Participants"
        ],
        "description": "IDs identify participants; the same email may occur on multiple records.",
        "responses": {
          "201": {
            "description": "Resource created; body includes its ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Participant"
                },
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "name": "Sam Lee",
                  "email": "sam@example.com"
                }
              }
            }
          },
          "400": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "invalidFields": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request contains invalid fields.",
                      "instance": "/api/participants",
                      "code": "INVALID_REQUEST",
                      "fieldErrors": [
                        {
                          "field": "email",
                          "message": "must be a valid email"
                        }
                      ]
                    }
                  },
                  "malformedJson": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request body must contain valid JSON.",
                      "instance": "/api/participants",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/participants",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParticipant"
              },
              "example": {
                "name": "Sam Lee",
                "email": "sam@example.com"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateCoach": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "Required nonblank name. ASCII whitespace is normalized; for example, \"  Alex   Rivera  \" returns \"Alex Rivera\".",
            "example": "  Alex   Rivera  "
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "pattern": "\\S",
            "example": "alex@example.com"
          }
        }
      },
      "Coach": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "Coach name with surrounding and repeated ASCII whitespace normalized to single spaces. Public field remains unchanged during internal name expansion.",
            "example": "Alex Rivera"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "pattern": "\\S",
            "example": "alex@example.com"
          }
        }
      },
      "CreateParticipant": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "Must contain a non-whitespace character.",
            "example": "Sam Lee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "pattern": "\\S",
            "example": "sam@example.com"
          }
        }
      },
      "Participant": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "22222222-2222-4222-8222-222222222222"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "Must contain a non-whitespace character.",
            "example": "Sam Lee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "pattern": "\\S",
            "example": "sam@example.com"
          }
        }
      },
...
      "ApiProblem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Request contains invalid fields."
          },
          "instance": {
            "type": "string",
            "format": "uri-reference",
            "example": "/api/coaches"
          },
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "COACH_NOT_FOUND",
              "PARTICIPANT_NOT_FOUND",
              "SESSION_NOT_FOUND",
              "REGISTRATION_NOT_FOUND",
              "COACH_OVERLAP",
              "DUPLICATE_REGISTRATION",
              "SESSION_FULL",
              "SESSION_HAS_REGISTRATIONS",
              "INTERNAL_ERROR"
            ],
            "example": "INVALID_REQUEST"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "description": "Present only for body field validation failures; sorted, deduplicated and omitted when empty."
          }
        }
      }
    }
  },
...
}